refactored clientDashboard.js, added cache to getProviders, reduced threshold to remove zoom out error, added search dropdown results, TODO: styling on clientdashboardsearchbar, server.js sometimes stops working... not sure why
