+++ client/src/css/carefinder.css
@@ -226,7 +226,6 @@ body#CF .providerButton button {
   background: black;
   padding: 60px;
 }
-
 p.amenities {
   margin-top: 20px;
 }
