+++ client/src/pages/ClientDashboard.js
@@ -1,34 +1,15 @@
 import React, { useState, useEffect, useRef } from 'react';
-import { Link } from 'react-router-dom';
-import {
-  GoogleMap,
-  useLoadScript,
-  Marker,
-  InfoWindowF,
-} from '@react-google-maps/api';
 import TopNav from '../components/TopNav';
 import Footer from '../components/Footer';
 import axios from 'axios';
-import ProviderCard from '../components/ProviderCard';
 import { debounce } from 'lodash';
-import { useLocation } from 'react-router-dom';
+import { useLocation, useNavigate } from 'react-router-dom';
 import { formatName } from '../utils';
 import mapStyles from '../mapStyles.json';
 import SurveyModal from '../components/SurveyModal';
 import { useAuth } from '../contexts/AuthContext';
-/**
- * TODO:  This page accepts a new param in the URL, sent here from the home page. Sample is:
- * /client-dashboard?refLookup=Seattle%2C%20WA
- * ...So, we need to grab that "refLookup" (if present), parse it as necessary, and then
- * pass it to the script below so that people can see their search immediately.
- * I imported useLocation, above to grab it, and then isolated it below as 'refLookup'
- * But I'm not sure how to best pass it to the page as a default location.
- * I think maybe we'd use a useEffect and, if the refLookup is present, we lookup its latitude/longitude using google Geocoder, and then set the lat/long as startingpoint instead of the existing default starting point. Since @sasha is most familiar here, I thought I'd leave it for you.
- *
- *
- * ALSO...
- * Anyone can now search this page and see results (no login required.) BUT, Micah wants it so that, when the provider card comes up, you have to be logged in to be able to continue (e.g., messsage the provider or reserve a room). I wasn't sure what the final flow is, but I think basically if they're not logged in, then the reserve room and message buttons would maybe either just go to the signup / login page, or maybe flash an alert on screen first to tell people they need to join (for free) first before continuing. I didn't make this change as I wasn't sure what the final flow would be.
- */
+import ClientDashboardSearchBar from '../components/ClientDashboardSearchBar';
+import ClientDashboardMap from '../components/ClientDashboardMap';
 
 export default function ClientDashboard() {
   const [searchQuery, setSearchQuery] = useState('');
@@ -37,7 +18,6 @@ export default function ClientDashboard() {
   const [selectedMarker, setSelectedMarker] = useState(null); // State to track selected marker
   const mapRef = useRef(null); // useRef to store the map instance
   const boundsRef = useRef(null); // useRef to store the boundaries
-  const providerListRef = useRef(null);
   const [isBoundsChanging, setIsBoundsChanging] = useState(false);
   const [initialLoad, setInitialLoad] = useState(true);
   const [zip, setZip] = useState(null);
@@ -47,17 +27,29 @@ export default function ClientDashboard() {
   const [surveyModalOpen, setSurveyModalOpen] = useState(false);
   const [hasSurvey, setHasSurvey] = useState(false);
   const { currentUser } = useAuth();
-
+  const navigate = useNavigate();
   let delayDuration = 500;
 
   const location = useLocation();
   const searchParams = new URLSearchParams(location.search);
   const refLookup = searchParams.get('refLookup');
 
+  useEffect(() => {
+    return () => {
+      debouncedGetProviders.current.cancel();
+    };
+  }, []);
+
+  useEffect(() => {
+    if (mapBounds) {
+      debouncedGetProviders.current(); // Only runs when `mapBounds` updates
+    }
+  }, [mapBounds]);
+
   // ADDED THIS TO GRAB a posible "refLookup" passed here.
   // Need to make this autoload on the page somehow, if present and valic
   // const location = useLocation();
-  // const searchParams = new URLSearchParams(location.search);
+  // const searchParams = new URLSearchParams(location.search);a
   // const refLookup = searchParams.get('refLookup');
 
   const getProvidersFromBounds = () => {
@@ -72,6 +64,7 @@ export default function ClientDashboard() {
         },
         center: startingPosition.current,
         radius: radiusRef.current,
+        zoomLevel: zoomRef.current,
       })
       .then((response) => {
         setProviders(response.data.providers);
@@ -108,32 +101,6 @@ export default function ClientDashboard() {
 
   const startingPosition = useRef({ lat: 47.7543, lng: -122.1635 });
 
-  const { isLoaded } = useLoadScript({
-    googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAPS_API_KEY,
-  });
-
-  const mapOptions = {
-    minZoom: 12,
-    styles: mapStyles,
-    disableDefaultUI: true,
-  };
-
-  const pinkMarkerIcon = {
-    url: 'selected_map_icon.png',
-    scaledSize: {
-      width: 50,
-      height: 50,
-    },
-  };
-
-  const mutedMarkerIcon = {
-    url: 'non_selected_icon.png',
-    scaledSize: {
-      width: 30,
-      height: 30,
-    },
-  };
-
   const radiusRef = useRef(3200);
   const zoomRef = useRef(13);
 
@@ -154,10 +121,6 @@ export default function ClientDashboard() {
       });
   }, []);
 
-  const handleMapLoad = (map) => {
-    mapRef.current = map; // Store the map instance in the ref
-  };
-
   useEffect(() => {
     if (initialLoad) {
       if (refLookup) {
@@ -170,6 +133,14 @@ export default function ClientDashboard() {
     }
     // eslint-disable-next-line
   }, []);
+
+  const handleFindMatches = () => {
+    if (hasSurvey) {
+      handleSearch();
+    } else {
+      navigate('/survey');
+    }
+  };
 
   const handleSearch = async (refQuery) => {
     let query;
@@ -189,6 +160,7 @@ export default function ClientDashboard() {
             address: query,
           }
         );
+        console.log('response', response);
         const { lat, lng } = response.data.address;
         startingPosition.current = { lat: lat, lng: lng };
         mapRef.current.panTo(startingPosition.current);
@@ -214,25 +186,6 @@ export default function ClientDashboard() {
     }
   };
 
-  const panToCity = (city) => {
-    if (mapRef.current) {
-      const cityCoords = { lat: city.lat, lng: city.lng };
-      startingPosition.current = cityCoords;
-      mapRef.current.panTo(cityCoords);
-      // setMapBounds(mapRef.current.getBounds()); // Manually set bounds
-      const fixedBounds = calculateFixedBounds(cityCoords, radiusRef.current); // 5000 meters (5 km) radius, adjust as needed
-      const newBounds = new window.google.maps.LatLngBounds(
-        new window.google.maps.LatLng(fixedBounds.south, fixedBounds.west),
-        new window.google.maps.LatLng(fixedBounds.north, fixedBounds.east)
-      );
-
-      boundsRef.current = newBounds;
-      // handleMapBoundsChanged();
-      setSelectedMarker(null);
-      getProvidersFromBounds();
-    }
-  };
-
   const calculateFixedBounds = (center, radiusInMeters) => {
     const earthRadius = 6378137; // Earth's radius in meters
     const latDelta = (radiusInMeters / earthRadius) * (180 / Math.PI);
@@ -248,48 +201,13 @@ export default function ClientDashboard() {
     };
   };
 
-  // const scrollToProvider = (providerId) => {
-  // 	if (providerListRef.current) {
-  //     console.log('jaks')
-  // 		const providerElement = providerListRef.current.querySelector(
-  // 			`#provider-${providerId}`
-  // 		);
-  // 		if (providerElement) {
-  //       console.log('asdfad')
-  // 			providerElement.scrollIntoView({ behavior: "smooth", block: "center" });
-  // 		}
-  // 	}
-  // };
-
-  const scrollToProvider = (licenseNumber) => {
-    const providerElement = providerListRef.current.querySelector(
-      `[data-license-number="${licenseNumber}"]`
-    );
-    if (providerElement) {
-      console.log(providerElement);
-      providerElement.scrollIntoView({ behavior: 'smooth', block: 'center' });
-    }
-  };
-
-  const handleProviderSelect = (provider) => {
-    // Scroll to the selected provider in the map
-    if (mapRef.current) {
-      mapRef.current.panTo(provider.position);
-      startingPosition.current = {
-        lat: provider.position.lat,
-        lng: provider.position.lng,
-      };
-      setSelectedMarker(provider); // Open InfoWindow for the selected provider
-    }
-  };
-
   const handleMapBoundsChanged = debounce(() => {
     if (mapRef.current) {
       const newBounds = mapRef.current.getBounds();
       const newCenter = newBounds.getCenter();
       const latDiff = Math.abs(startingPosition.current.lat - newCenter.lat());
       const lngDiff = Math.abs(startingPosition.current.lng - newCenter.lng());
-      const threshold = 0.00001; // Adjust the threshold value as needed
+      const threshold = 0.001; // Adjust the threshold value as needed
 
       if (initialLoad || latDiff > threshold || lngDiff > threshold) {
         boundsRef.current = newBounds;
@@ -298,6 +216,7 @@ export default function ClientDashboard() {
         } else if (mapRef.current.getZoom() < zoomRef.current) {
           radiusRef.current *= 2;
         }
+
         zoomRef.current = mapRef.current.getZoom();
         startingPosition.current = {
           lat: newCenter.lat(),
@@ -316,239 +235,35 @@ export default function ClientDashboard() {
     }
   }, [isBoundsChanging, mapBounds]);
 
-  const handleMarkerClick = (marker) => {
-    setSelectedMarker(marker);
-    scrollToProvider(marker.LicenseNumber);
-  };
-
-  console.log('providers', providers);
-
   return (
     <>
       <TopNav />
-      <div className="CFblackBackground">
-        <div
-          className="contentContainer clientSearchBar"
-          style={{
-            display: 'flex',
-            justifyContent: 'center',
-            alignItems: 'center',
-            padding: '20px',
-          }}
-        >
-          <input
-            id="clientSearch"
-            name="yyy"
-            placeholder="Search city, zip code etc."
-            value={searchQuery}
-            onChange={(e) => setSearchQuery(e.target.value)}
-            onKeyDown={(e) => {
-              if (e.key === 'Enter') {
-                handleSearch();
-              }
-            }}
-            style={{
-              padding: '20px 20px 20px 20px',
-              borderRadius: '30px',
-              width: '50%',
-              marginRight: '10px',
-              position: 'relative',
-              top: 6,
-              fontSize: '20px',
-            }}
-          />
-          <button
-            id="findMatches"
-            type="button"
-            className="btn"
-            onClick={() => handleSearch()}
-            style={{
-              padding: '10px 20px',
-              borderRadius: '10px',
-              backgroundColor: '#FFA500',
-              color: 'white',
-              marginRight: '10px',
-            }}
-          >
-            Find a Match
-          </button>
-          <Link
-            to="/survey"
-            id="takeSurvey"
-            type="button"
-            className="btn"
-            style={{
-              padding: '10px 20px',
-              borderRadius: '10px',
-              backgroundColor: '#FF69B4',
-              color: 'white',
-            }}
-          >
-            Contact us
-          </Link>
-        </div>
-      </div>
+      <ClientDashboardSearchBar
+        searchQuery={searchQuery}
+        setSearchQuery={setSearchQuery}
+        handleSearch={handleSearch}
+        handleFindMatches={handleFindMatches}
+        errorMessage={errorMessage}
+      />
 
-      {errorMessage && (
-        <div
-          style={{
-            color: 'red',
-            marginTop: '10px',
-            textAlign: 'center',
-            width: '100%',
-          }}
-        >
-          {errorMessage}
-        </div>
-      )}
-
-      <div className="CFblackBackground">
-        <div className="contentContainer" style={{ padding: '20px' }}>
-          <div style={{ textAlign: 'center', marginBottom: '20px' }}>
-            <h3>{city ? city : 'Woodinville'}, WA</h3>
-            <h6>Licensed AFH care providers in {zip ? zip : '98072'}</h6>
-          </div>
-          <div className="providersRow" style={{ display: 'flex' }}>
-            <div
-              className="googleMapsContainer"
-              style={{ flex: 1, paddingRight: '20px' }}
-            >
-              {isLoaded ? (
-                <GoogleMap
-                  zoom={zoomRef.current}
-                  center={startingPosition.current}
-                  mapContainerStyle={{
-                    width: '100%',
-                    height: '500px',
-                    borderRadius: 10,
-                  }}
-                  onLoad={handleMapLoad}
-                  onBoundsChanged={handleMapBoundsChanged}
-                  options={mapOptions}
-                >
-                  {providers &&
-                    providers.length > 0 &&
-                    providers.map((provider) => (
-                      <Marker
-                        key={provider.id}
-                        position={provider.position}
-                        onClick={() => handleMarkerClick(provider)}
-                        className="h-20 w-20"
-                        icon={
-                          selectedMarker === provider
-                            ? pinkMarkerIcon
-                            : mutedMarkerIcon
-                        }
-                      />
-                    ))}
-                  {selectedMarker && (
-                    <InfoWindowF
-                      position={selectedMarker.position}
-                      onCloseClick={() => setSelectedMarker(null)}
-                    >
-                      <div className="relative text-black flex flex-col max-w-[15rem]">
-                        {/* Image Container with Close Button */}
-                        <div className="relative w-full">
-                          <img
-                            src={selectedMarker.listingsData.homePhotos[0]}
-                            alt="Home Photo"
-                            className="w-full h-32 object-cover"
-                          />
-                          {/* Close Button positioned absolutely */}
-                          <button
-                            className="absolute top-2 right-2  hover:bg-white  bg-transparent rounded-full p-1 shadow-md text-black"
-                            onClick={() => setSelectedMarker(null)}
-                          >
-                            ✕
-                          </button>
-                        </div>
-
-                        {/* Content Section */}
-                        <div className="p-2">
-                          <h5 className="font-semibold">
-                            {selectedMarker.FacilityName}
-                          </h5>
-                          <div>{selectedMarker.LocationAddress}</div>
-                          <div>{`${selectedMarker.LocationCity}, ${selectedMarker.LocationState} ${selectedMarker.LocationZipCode}`}</div>
-                          <div className="mt-2 text-center">
-                            {
-                              selectedMarker.roomsData.filter(
-                                (room) => room.isAvailable
-                              ).length
-                            }{' '}
-                            rooms available
-                          </div>
-                        </div>
-                      </div>
-                    </InfoWindowF>
-                  )}
-                </GoogleMap>
-              ) : (
-                <div>Loading ...</div>
-              )}
-              <div>
-                <h6
-                  style={{
-                    textAlign: 'left',
-                    marginTop: '4px',
-                    marginBottom: '2px',
-                  }}
-                >
-                  Neighboring cities near {zip}
-                </h6>
-                <ul
-                  style={{
-                    display: 'flex',
-                    gap: '20px',
-                    justifyContent: 'left',
-                    paddingLeft: 0,
-                  }}
-                >
-                  {nearbyBigCities &&
-                    nearbyBigCities.map((city) => (
-                      <li
-                        key={city.name}
-                        onClick={() => panToCity(city)}
-                        style={{
-                          listStyle: 'none',
-                          textAlign: 'left',
-                          cursor: 'pointer',
-                        }}
-                      >
-                        {/* <img src={city.image} alt={city.name} style={{ width: "50px", height: '50px', objectFit: "cover", borderRadius: "5px", margin: '0 auto' }} /> */}
-                        <div>{city.name}</div>
-                      </li>
-                    ))}
-                </ul>
-              </div>
-            </div>
-
-            <div
-              className="pcards_right CFblackBackground"
-              style={{
-                flex: 1,
-                maxHeight: '500px',
-                overflowY: 'auto',
-              }}
-              ref={providerListRef}
-            >
-              {providers && providers.length > 0 ? (
-                providers.map((provider) => (
-                  <ProviderCard
-                    key={provider.LicenseNumber}
-                    provider={provider}
-                    onClick={() => handleProviderSelect(provider)}
-                    hasSurvey={hasSurvey}
-                    setSurveyModalOpen={setSurveyModalOpen}
-                  />
-                ))
-              ) : (
-                <div>No providers found</div>
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
+      <ClientDashboardMap
+        providers={providers}
+        selectedMarker={selectedMarker}
+        setSelectedMarker={setSelectedMarker}
+        zip={zip}
+        city={city}
+        zoomRef={zoomRef}
+        startingPosition={startingPosition}
+        handleMapBoundsChanged={handleMapBoundsChanged}
+        nearbyBigCities={nearbyBigCities}
+        hasSurvey={hasSurvey}
+        setSurveyModalOpen={setSurveyModalOpen}
+        mapRef={mapRef}
+        boundsRef={boundsRef}
+        radiusRef={radiusRef}
+        mapStyles={mapStyles}
+        getProvidersFromBounds={getProvidersFromBounds}
+      />
       <SurveyModal
         showModal={surveyModalOpen}
         handleCloseModal={() => setSurveyModalOpen(false)}
